+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha.11] - 2023-05-17
+- add: expose `setDracoDecoderLocation` and `setKTX2TranscoderLocation`
+- fix: allow using draco decoder and ktx2 transcoder from local filepath
+
 ## [1.0.0-alpha.10] - 2023-05-07
 - fix: progressive assets are now only updated during the main canvas render call and not e.g. when rendering to a texture